requirements.txt
# ---- Núcleo FastAPI/Starlette/WSGI-ASGI ----
fastapi==0.95.2
uvicorn==0.22.0
gunicorn==21.2.0
starlette==0.27.0
jinja2
python-multipart==0.0.6

# ---- Flask (AI Detect) ----
flask==3.0.0
python-dotenv==1.0.0

# ---- OpenAI + audio utilidades (Detect/Tutor) ----
openai==0.27.8
websockets==11.0.3
gTTS==2.3.2
pydub==0.25.1
SpeechRecognition==3.10.0
PyPDF2==3.0.1
pytesseract==0.3.10
Pillow==10.0.0

# ---- Visión/YOLO (Seguridad) ----
opencv-python
ultralytics
numpy